Converge job service
