+++ src/test/groovy/io/seqera/wave/service/blob/impl/DockerTransferStrategyTest.groovy
@@ -19,14 +19,13 @@
 package io.seqera.wave.service.blob.impl
 
 import spock.lang.Specification
-import spock.lang.Unroll
 
 import io.seqera.wave.configuration.BlobCacheConfig
 /**
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
-class DockerJobStrategyTest extends Specification {
+class DockerTransferStrategyTest extends Specification {
 
     def 'should create transfer cli' () {
         given:
@@ -65,17 +64,4 @@ class DockerJobStrategyTest extends Specification {
         result.redirectErrorStream()
     }
 
-    @Unroll
-    def 'should parse state string' () {
-        expect:
-        DockerTransferStrategy.State.parse(STATE) == EXPECTED
-
-        where:
-        STATE               | EXPECTED
-        'running'           | new DockerTransferStrategy.State('running')
-        'exited'            | new DockerTransferStrategy.State('exited')
-        'exited,'           | new DockerTransferStrategy.State('exited')
-        'exited,0'          | new DockerTransferStrategy.State('exited', 0)
-        'exited,10'         | new DockerTransferStrategy.State('exited', 10)
-    }
 }
